pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Kalil <kalilgarcia38@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.112.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.2.1"}
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 100
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev fast_zero/app.py'
runip = 'task run --host 192.168.1.54'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
